% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MEgbm.R
\name{MEgbm}
\alias{MEgbm}
\title{MEgbm}
\usage{
MEgbm(X, Y, groups = NULL, rand.vars = "1", para = NULL,
  gbm.dist = "bernoulli", lme.family = binomial, tol = 1e-05,
  max.iter = 100, include.RE = TRUE, verbose = FALSE,
  likelihoodCheck = TRUE, type = "prob", ...)

MEgbm(X, Y, groups = NULL, rand.vars = "1", para = NULL,
  gbm.dist = "bernoulli", lme.family = binomial, tol = 1e-05,
  max.iter = 100, include.RE = TRUE, verbose = FALSE,
  likelihoodCheck = TRUE, type = "prob", ...)
}
\arguments{
\item{X}{data.frame with predictors}

\item{Y}{binary response vector}

\item{groups}{character name of the column containing the group identifier}

\item{rand.vars}{random effect variables}

\item{para}{named list of gbm training parameters}

\item{gbm.dist}{gbm loss function}

\item{lme.family}{glmer control}

\item{tol}{convergence tolerance}

\item{max.iter}{maximum number of iterations}

\item{include.RE}{(logical) to include random effect Zb as predictor in gbm?}

\item{verbose}{verbose for lme4}

\item{likelihoodCheck}{(logical) to use log likelihood of glmer to check for convergence?}

\item{type}{of predictions of gbm to pass to lme4 as population estimates (these will be used as offset)}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
An object of class MEgbm; a list with items
\item{gbmfit}{fitted gbm model}
\item{glmer.fit}{fitted mixed effect logistic regression model}
\item{logLik}{log likelihood of mixed effect logistic regression}
\item{random.effects}{random effect parameter estimates}
\item{boost.form}{gbm formula for fitted model}
\item{glmer.form}{lmer4 formula}
\item{glmer.CI}{estimates of mixed effect logistic regression with
    approximate confidence intervals on the logit scale. More accurate values
    can be obtained by bootstrap}
\item{fitted.probs}{fitted probabilites for final model}
\item{fitted.class}{fitted class labels for final model}
\item{train.perf}{various performance measures for final model on training set}
\item{threshold}{classification cut-off}
}
\description{
Trains a Mixed Effect Boosted regression and
 classification models for longitudinal continuous, binary and count data




}
\examples{
\dontrun{
set.seed(12345)
mod <- MEgbm(form, rand.form, data))

}
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}

