% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.MEgbm.R
\name{mkNewReTrms}
\alias{mkNewReTrms}
\title{Make new random effect terms from specified object and new data,
possibly omitting some random effect terms}
\usage{
mkNewReTrms(object, newdata, re.form = NULL, allow.new.levels = FALSE)
}
\arguments{
\item{object}{fitted model object}

\item{newdata}{(optional) data frame containing new data}

\item{re.form,allow.new.levels}{formula specifying random effect terms to include (NULL=all, ~0)}
}
\description{
Make new random effect terms from specified object and new data,
possibly omitting some random effect terms
}
\note{
Hidden; _only_ used (twice) in this file
}

