% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.MEglmTree.R
\docType{data}
\name{Plot.MEglmTree}
\alias{.plot_node}
\alias{NodeBarplot}
\alias{NodeBoxplot}
\alias{Plot.MEglmTree}
\alias{heat.tree}
\alias{plot.MECTree}
\alias{plot.MEglmTree}
\alias{print.MECTree}
\title{Plot method for MEglmTree object}
\format{\preformatted{function (obj, fitted = TRUE, col = "black", fill = "lightgray", width = 0.5, yscale = NULL, ylines = 3, cex = 0.5, id = TRUE, mainlab = NULL, gp = gpar(), 
    ...)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 251 16 362 1 16 1 251 362
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xd1612a0> 
 - attr(*, "class")= chr "grapcon_generator"
}}
\usage{
\method{plot}{MEglmTree}(x, main = NULL, terminal_panel = node_terminal,
  tp_args = list(), inner_panel = node_inner, ip_args = list(),
  edge_panel = edge_simple, ep_args = list(), drop_terminal = FALSE,
  tnex = 1, newpage = TRUE, pop = TRUE, gp = gpar(), ...)

\method{plot}{MECTree}(x, ...)

\method{print}{MECTree}(x, ...)

heat.tree(tree, low.is.green = FALSE, ...)

.plot_node(node, xlim, ylim, nx, ny, terminal_panel, inner_panel, edge_panel,
  tnex = 2, drop_terminal = TRUE, debug = FALSE)

NodeBoxplot

NodeBarplot
}
\arguments{
\item{x}{fited MEglmTree or MECtree object.}

\item{main,ep_args}{see partykid}

\item{terminal_panel}{see partykit}

\item{tp_args,ip_args,xlim,ylim}{see partykit}

\item{inner_panel}{see partykit}

\item{edge_panel}{see partykit}

\item{drop_terminal}{see partykit}

\item{tnex}{see partykit}

\item{newpage,pop,gp,tree,low.is.green}{see partykit}

\item{node}{see partykit}

\item{nx,ny}{see partykit}

\item{debug}{see partykit}

\item{\dots}{Further arguments passed to partykit}
}
\description{
Take a MEglmTree and plot the barplot of fitted classes or boxplot of
fitted probability










}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}
\keyword{datasets}

