% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{safeDeparse}
\alias{safeDeparse}
\title{deparse(.) returning \bold{one} string}
\usage{
safeDeparse(x, collapse = " ")
}
\arguments{
\item{x,collapse}{character vector and how to combine them}
}
\description{
deparse(.) returning \bold{one} string
}
\note{
Protects against the possibility that results from deparse() will be
      split after 'width.cutoff' (by default 60, maximally 500)
}

