% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils}
\alias{LongiLagSplit}
\alias{Performance.measures}
\alias{Utils}
\alias{cprod}
\alias{insampleTrainTestSplit}
\alias{insampleTrainTestSplit.mtl}
\alias{levelfun}
\alias{lhs.form}
\alias{opt.thresh}
\alias{reOnly}
\alias{rhs.form}
\alias{sigmoid}
\alias{trajTrainTestSplit}
\title{Utility functions.}
\usage{
insampleTrainTestSplit(dat, id, order.vars = "id", cut.off = 4,
  propo = 0.25, no.last = 3)

insampleTrainTestSplit.mtl(dat, id, resp.vars, rhs.vars, cut.off = 4,
  propo = 0.25, no.last = 3)

cprod(x, y = NULL)

reOnly(f)

levelfun(x, nl.n, allow.new.levels = FALSE)

trajTrainTestSplit(dat, id, rhs.vars, resp.vars, order.vars = "time",
  lag = 1)

LongiLagSplit(dat, id, rhs.vars, resp.vars, order.vars = "time", lag = 1)

Performance.measures(pred, obs, threshold = NULL)

sigmoid(x)

lhs.form(formula)

rhs.form(formula)

opt.thresh(pred, obs)
}
\arguments{
\item{dat}{data.frame}

\item{id}{group id}

\item{order.vars}{variables to order longitudinal data}

\item{cut.off,propo}{cut-off and proportions}

\item{no.last}{number of tail observations that goes into test set}

\item{resp.vars,rhs.vars}{response and vector of predictor names}

\item{x}{a random effect (i.e., data frame with rows equal to levels, columns equal to terms}

\item{y,f,nl.n}{numeric or character vector depending on usage}

\item{allow.new.levels}{logical}

\item{lag}{lag in train test split}

\item{pred,obs}{predicted probabilities and true class}

\item{threshold}{threshold}

\item{formula}{formula}

\item{\dots}{Further arguments passed to or from other methods.}

\item{n}{vector of new levels}
}
\description{
Some useful miscellaneous functions: plot plot.MECTree plot the
  tree part of MECTree object; heat.tree plot a sort of heat map decision trees
















}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}

